requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
pydantic
requests
sqlalchemy
psycopg2-binary
python-dotenv